feat(staff): tag the revenue tables with Stripe's logo instead of a column
Both tables carried a column whose only content was the word "Stripe" on
every row. The logo now sits at the edge of the team's own cell, as it does on
the trial pipeline — a mark on the team rather than a column of its own — and
the width it frees goes to the figures.

The component moves out of the trials page into `stripe.tsx`, next to the URL
it is built from, so the third page to want it does not copy it a second time.
